_"Goodnight, my love."_ - he said while typing this commit message.
_"You know, it pains me to do this to you, but this isn't healthy for the both of us"_ - he continued.
They were in a constant battle. He was trying to make the best of his time, and it was just breaking on every update.

Him being @NickAcPT and it being winit. It is insufferable.
He is tired of bumping a version number and having to fix a bunch of stuff.
**He is tired of having to fix a bunch of stuff.**

Thus, he did what he had to do. He broke up with winit.
He is now with just wgpu (in nmsr-aas). It is a healthier relationship.

Not having to deal with windowing stuff is a relief.
He can now focus on the rendering part of the relationship.

He is happy.

He is free.

He is NickAcPT.

---

Like seriously, please winit, STOP BREAKING EVERY MINOR VERSION BUMP.
Like, yeah, sure, I'm aware that the major version is still 0, but come on, this is just ridiculous.
I've grown tired of it.